requirements.txt
streamlit 
pdfplumber
google-generativeai 
python-dotenv